Gewichte in den Funktionen und dem Graphen von double zu int geändert

+++ Einführungsaufgabe/src/steinergraph.cpp
@@ -93,7 +93,7 @@ namespace
     * Checking for lines belonging to the graph-section.
     * The output-number tells which kind of information was in the line given.
     */
-   int graph_section(const std::string &_line, int &_num_nodes, int &_num_edges, int &_head, int &_tail, double &_weight, int &_edge_counter)
+   int graph_section(const std::string &_line, int &_num_nodes, int &_num_edges, int &_head, int &_tail, int &_weight, int &_edge_counter)
    {
       std::stringstream ss(_line);
       std::string _keyword = "";
@@ -173,7 +173,7 @@ namespace
 }
 
 const SteinerGraph::NodeId SteinerGraph::invalid_node = -1;
-const double SteinerGraph::infinite_weight = std::numeric_limits<double>::max();
+const int SteinerGraph::infinite_weight = std::numeric_limits<int>::max();
 const int SteinerGraph::infinite_distance = -1;
 
 void SteinerGraph::add_nodes(NodeId num_new_nodes)
@@ -195,11 +195,11 @@ void SteinerGraph::Node::set_terminal()
    _terminal = true;
 }
 
-SteinerGraph::Neighbor::Neighbor(SteinerGraph::NodeId n, double w) : _id(n), _edge_weight(w) {}
+SteinerGraph::Neighbor::Neighbor(SteinerGraph::NodeId n, int w) : _id(n), _edge_weight(w) {}
 
 SteinerGraph::SteinerGraph(NodeId num) : _nodes(num) {}
 
-void SteinerGraph::add_edge(NodeId tail, NodeId head, double weight)
+void SteinerGraph::add_edge(NodeId tail, NodeId head, int weight)
 {
    if (tail >= num_nodes() or tail < 0 or head >= num_nodes() or head < 0)
    {
@@ -209,7 +209,7 @@ void SteinerGraph::add_edge(NodeId tail, NodeId head, double weight)
    _nodes[head].add_neighbor(tail, weight);
 }
 
-void SteinerGraph::Node::add_neighbor(SteinerGraph::NodeId nodeid, double weight)
+void SteinerGraph::Node::add_neighbor(SteinerGraph::NodeId nodeid, int weight)
 {
    _neighbors.push_back(SteinerGraph::Neighbor(nodeid, weight));
 }
@@ -243,7 +243,7 @@ bool SteinerGraph::Node::is_terminal() const
    return _terminal;
 }
 
-double SteinerGraph::Neighbor::edge_weight() const
+int SteinerGraph::Neighbor::edge_weight() const
 {
    return _edge_weight;
 }
@@ -307,7 +307,7 @@ SteinerGraph::SteinerGraph(char const *filename) // Konstruktor der Klasse   -  
    int num_nodes = -1, num_edges = -1, num_terminals = -1; // different variables for the read routines
    int edge_counter = 0, terminal_counter = 0;
    int head = -1, tail = -1;
-   double weight = 1.0;
+   int weight = 1;
    std::string keyword = "";
 
    while (std::getline(file, line)) // while-loop that reads the file line by line
